refactor FlashcardMemoryGame to use mode components
